+++ app/js/main.js
@@ -34,44 +34,69 @@ function showError(text) {
   }
 }
 
+function newUnitColumn(unitID) {
+  var cardHeader = crel("h3", {
+    "class": "card-header"
+  }, unitID);
+
+  var card =  crel("div", {
+    "class": "card " + unitID
+  }, cardHeader /* , _cardBlock_ */ );
+
+  return card;
+}
+
+function newCardBlock(classElement) {
+  var className = classElement.getAttribute("className");
+  var cardTitle = crel("h4", {
+    "class": "card-title"
+  }, className);
+
+  var classType = classElement.getAttribute("classType");
+  var cardBlock = crel("div", {
+    "class": "card-block " + classType
+  }, cardTitle);
+
+  return cardBlock;
+}
+
 /**
  * Generate a nice little output of the classes the user has selected
  * so they can be ready for registration day
+ * TODO Make sure this function is only called once per update (check refresh)
  */
 function generateClassOutput() {
-  // TODO Replace this class overview with Bootstrap 4 Cards
-  $(".unitOverview").html(""); // Clear what was there before
+  // TODO Find a way to organise cards left-to-right to prevent empty columns
 
-  // Create headings and containers for selected class data
-  var unitNames = $(".unit-name");
-  var len = unitNames.length, i = 0;
-  for (i; i < len; i++) {
-    unitID = unitNames[i].textContent;
-    var unitContainer = crel("div",
-      {"id": unitID}, // Unit Container
-      crel("h3", unitID), // Unit Heading
-      crel("div", {"class": "selected_classes"}) // Class Container
-    );
-    $(".unitOverview").append(unitContainer);
-  }
+  var cardRow = $(".unitOverview");
+  cardRow.empty(); // Clear what was there before
 
-  // Add the classes into the selected classes section we created before
   var selectedClasses = $(".class:selected");
-  len = selectedClasses.length;
-  i = 0;
+  var len = selectedClasses.length, i = 0;
   for (i; i < len; i++) {
-    var overview = getClassOverview(selectedClasses[i]);
-
-    // Check for missing data
-    if (overview.indexOf("undefined") > -1) {
-      showError("Old or broken class data has been detected! " +
-        "This may be caused by an update, or by invalid HTML editing. " +
-        "Note that you should refresh the page if you fix manually.");
+    // Check if there is a card for this unit
+    var unitID = selectedClasses[i].getAttribute("unitID");
+    var unitCard = $(".card." + unitID);
+    if (unitCard.length < 1) {
+      // Create the unit card
+      unitCard = newUnitColumn(unitID);
+      cardRow.append(unitCard);
     }
 
-    // Append selected class data to the container with the same unit ID
-    var unitID = $(selectedClasses[i]).parents().eq(2).children(".unit-name").text();
-    $("#" + unitID + ">.selected_classes").append(overview);
+    // Check if there is a type header for this class
+    var classType = selectedClasses[i].getAttribute("classType");
+    var typeBlock = $(unitCard).find(".card-block." + classType);
+    if (typeBlock.length < 1) {
+      // Create new type block
+      typeBlock = newCardBlock(selectedClasses[i]);
+      unitCard.append(typeBlock);
+    }
+
+    // Get the class-text
+    var cardText = crel("p", {
+      "class": "card-text"
+    }, getClassOverview(selectedClasses[i]));
+    typeBlock.append(cardText);
   }
 }